re-design the database layout, each database only contain one translation unit. the dbname will be clang_[userdefine]_[prefix-

+++ sspsql.c
@@ -4,8 +4,14 @@
 #define DBUSER  "ssp"
 #define DBPWD   "ssp"
 
-#define QUERY(x,y) printf("[sql]:%s\n",y);if (mysql_query((x), (y))) printf("Error %u: %s\n", mysql_errno(x), mysql_error(x))
-#define HEADER "id BIGINT NOT NULL AUTO_INCREMENT, kind CHAR(64) NOT NULL, spell CHAR(128), type CHAR(64) NOT NULL, hash BIGINT, primary key (id)" 
+#define QUERY(x,y) printf("[sql]:%s\n",y);\
+    if (mysql_query((x), (y))) \
+    printf("Error %u: %s\n", mysql_errno(x), mysql_error(x))
+
+#define HEADER "id BIGINT NOT NULL AUTO_INCREMENT, \
+    kind CHAR(64) NOT NULL, spell CHAR(128), \
+    type CHAR(64) NOT NULL, hash BIGINT, primary key (id)" 
+
 #define HEAD "kind, spell, type, hash"
 
 const char*
@@ -35,12 +41,14 @@ sql_close(MYSQL *conn){
     mysql_close(conn);}
 
 void 
-sql_create_tbl(MYSQL *conn, char *tblname){ 
+sql_create_tbl(MYSQL *conn, char *tblname, unsigned int flag){ 
     char *query = malloc(256 * sizeof(char));
     if (conn == NULL){printf(" Connection Error\n");}
-    sprintf(query, "drop table IF EXISTS %s", tblname);
-    QUERY(conn, query);
-    sprintf(query, "create table %s (%s)", tblname, HEADER);
+    if(flag == LOST){
+        sprintf(query, "drop table IF EXISTS %s", tblname);
+        QUERY(conn, query);
+    }
+    sprintf(query, "create table IF NOT EXISTS %s (%s)", tblname, HEADER);
     QUERY(conn, query);
     free(query);
 }
